Fixed compatibility for ProtocolLib and InteractiveChat
