+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>me.ryzenfromfire</groupId>
     <artifactId>PressF</artifactId>
-    <version>1.1.1-SNAPSHOT</version>
+    <version>1.1.2</version>
     <packaging>jar</packaging>
 
     <name>PressF</name>
