+++ src/main/java/me/ryzenfromfire/pressf/PressF.java
@@ -1,7 +1,12 @@
 package me.ryzenfromfire.pressf;
 
+import com.comphenix.protocol.PacketType;
 import com.comphenix.protocol.ProtocolLibrary;
 import com.comphenix.protocol.ProtocolManager;
+import com.comphenix.protocol.events.ListenerPriority;
+import com.comphenix.protocol.events.PacketAdapter;
+import com.comphenix.protocol.events.PacketContainer;
+import com.comphenix.protocol.events.PacketEvent;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import net.kyori.adventure.text.minimessage.Template;
@@ -11,6 +16,7 @@ import org.bukkit.OfflinePlayer;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
+import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 
@@ -81,6 +87,31 @@ public final class PressF extends JavaPlugin {
         } else {
             getLogger().info("ProtocolLib not found.");
         }
+        if (protocolLibHook) {
+            ProtocolManager protocolManager = getProtocolManager();
+            if (protocolManager == null) {
+                getLogger().severe("ERROR: ProtocolLib Hook failed (null).");
+            } else {
+                protocolManager.addPacketListener(new PacketAdapter((Plugin) this,
+                        ListenerPriority.NORMAL,
+                        PacketType.Play.Client.CHAT) {
+                    @Override
+                    public void onPacketReceiving(PacketEvent event) {
+                        if (event.getPacketType() == PacketType.Play.Client.CHAT) {
+                            PacketContainer packet = event.getPacket();
+                            String message = packet.getStrings().read(0);
+                            if (message.equalsIgnoreCase("F")) {
+                                event.setCancelled(true);
+                                Bukkit.getScheduler().runTask(this.plugin, () -> Bukkit.dispatchCommand(event.getPlayer(), "pressf"));
+                            } else {
+                                events.setLastMessenger(event.getPlayer());
+                                events.setLastMessageTime(System.currentTimeMillis());
+                            }
+                        }
+                    }
+                });
+            }
+        } //end PL Hook
     }
 
     @Override
